requirements.txt
Django>=5.1,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
drf-spectacular>=0.27,<0.28
django-filter>=24.1,<25.0
django-cors-headers>=4.3,<5.0
psycopg2-binary>=2.9,<3.0
qrcode[pil]>=7.4,<8.0
openpyxl>=3.1,<4.0
gunicorn>=22.0,<23.0
